Split up tests, increase coverage
