+++ tests/utils.ts
@@ -6,7 +6,6 @@ import {
   recordOrIdArray,
   recordOrNull,
   recordOrUndefined,
-  RecordWithConstructor,
   SimpleRecord,
 } from '../src/';
 
@@ -18,73 +17,6 @@ type ISimpleInterface = Readonly<{
 const MySimpleRecord = SimpleRecord<ISimpleInterface>({
   a: 5,
   b: '--',
-});
-
-describe('SimpleRecord', () => {
-  it('should provide defaults to an empty constructor', () => {
-    const instance = MySimpleRecord();
-    expect(instance.a).toBe(5);
-    expect(instance.b).toBe('--');
-  });
-
-  it('should mix valid data in over defaults', () => {
-    const instance = MySimpleRecord({
-      b: 'llama',
-    });
-    expect(instance.a).toBe(5);
-    expect(instance.b).toBe('llama');
-  });
-});
-
-interface IComplexInterfaceShared {
-  first: string;
-}
-type IComplexInterfaceInput = Readonly<
-  IComplexInterfaceShared & {
-    second: Partial<ISimpleInterface>;
-  }
->;
-type IComplexInterface = Readonly<
-  IComplexInterfaceShared & {
-    second: ISimpleInterface;
-  }
->;
-
-const MyComplexRecord = RecordWithConstructor<
-  IComplexInterfaceInput,
-  IComplexInterface
->(
-  {
-    first: 'first',
-    second: {},
-  },
-  input => {
-    return {
-      ...input,
-      second: MySimpleRecord(input.second),
-    };
-  }
-);
-
-describe('RecordWithConstructor', () => {
-  it('should deeply construct my items', () => {
-    const instance = MyComplexRecord();
-    expect(instance.first).toBe('first');
-    expect(instance.second.a).toBe(5);
-    expect(instance.second.b).toBe('--');
-  });
-
-  it('should provide correctly merged defaults', () => {
-    const instance = MyComplexRecord({
-      first: 'a',
-      second: {
-        a: 10,
-      },
-    });
-    expect(instance.first).toBe('a');
-    expect(instance.second.a).toBe(10);
-    expect(instance.second.b).toBe('--');
-  });
 });
 
 describe('recordOrUndefined', () => {
@@ -129,6 +61,11 @@ describe('recordArray', () => {
     expect(items[0].a).toBe(5);
     expect(items[1].a).toBe(5);
   });
+
+  it('should handle empty inputs', () => {
+    const items = recordArray(undefined, MySimpleRecord);
+    expect(items).toEqual([]);
+  });
 });
 
 describe('recordOrIdArray', () => {
@@ -168,5 +105,10 @@ describe('narrowToRecordArray', () => {
     expect(() => {
       narrowToRecordArray(['5', '6']);
     }).toThrow();
+  });
+
+  it('should handle empty arrays', () => {
+    const items = narrowToRecordArray([]);
+    expect(items).toEqual([]);
   });
 });
